feat(e2e): :arrow_up: migrate e2e app to Next 12.2
adapt to updated recommendation of `next/script` usage.
Scripts with `beforeInteractive` have to be placed in `_document` now:
https://nextjs.org/docs/basic-features/script#beforeinteractive

For inline scripts that need to load beforeInteractive, can be put
as directly as children of `<Head>` in _document and will automatically
hashed/nonced
