+++ apps/e2e/next.config.js
@@ -1,18 +1,8 @@
-const nextSafe = require("next-safe");
-
-const isDev = process.env.NODE_ENV !== "production";
-
-/** @type {import('next').NextConfig['headers']} */
-const headers = async () => [
-  {
-    source: "/:path*",
-    headers: nextSafe({ isDev }),
-  },
-];
 
 /** @type {import('next').NextConfig} */
 module.exports = {
   reactStrictMode: true,
+  crossOrigin: "anonymous",
   swcMinify: true,
   experimental: {
     workerThreads: true,
